--- src/routes/(private)/_scoped/projects/$project_id.lazy.tsx
@@ -1,54 +1,0 @@
-import { ProjectScopedSidebar } from "@/components/sidebar/app-sidebar";
-import {
-	Breadcrumb,
-	BreadcrumbItem,
-	BreadcrumbLink,
-	BreadcrumbList,
-	BreadcrumbPage,
-	BreadcrumbSeparator,
-} from "@/components/ui/breadcrumb";
-import { Separator } from "@/components/ui/separator";
-import {
-	SidebarInset,
-	SidebarProvider,
-	SidebarTrigger,
-} from "@/components/ui/sidebar";
-import { createLazyFileRoute, Outlet } from "@tanstack/react-router";
-
-export const Route = createLazyFileRoute(
-	"/(private)/_scoped/projects/$project_id"
-)({
-	component: RouteComponent,
-});
-
-function RouteComponent() {
-	return (
-		<SidebarProvider>
-			<ProjectScopedSidebar />
-			<SidebarInset className="overflow-x-hidden">
-				<header className="flex h-16 shrink-0 items-center gap-2">
-					<div className="flex items-center gap-2 px-4">
-						<SidebarTrigger className="-ml-1" />
-						<Separator orientation="vertical" className="mr-2 h-4" />
-						<Breadcrumb>
-							<BreadcrumbList>
-								<BreadcrumbItem className="hidden md:block">
-									<BreadcrumbLink href="#">
-										Building Your Application
-									</BreadcrumbLink>
-								</BreadcrumbItem>
-								<BreadcrumbSeparator className="hidden md:block" />
-								<BreadcrumbItem>
-									<BreadcrumbPage>Data Fetching</BreadcrumbPage>
-								</BreadcrumbItem>
-							</BreadcrumbList>
-						</Breadcrumb>
-					</div>
-				</header>
-				<div className="flex flex-1 flex-col gap-4 p-4 pt-0">
-					<Outlet />
-				</div>
-			</SidebarInset>
-		</SidebarProvider>
-	);
-}
