+++ src/routes/(private)/_scoped/projects/$project_id/index.tsx
@@ -1,6 +1,6 @@
-import { createLazyFileRoute } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createLazyFileRoute(
+export const Route = createFileRoute(
 	"/(private)/_scoped/projects/$project_id/"
 )({
 	component: RouteComponent,
