--- src/routes/(private)/_unscoped.lazy.tsx
@@ -1,52 +1,0 @@
-import { UnscopedSidebar } from "@/components/sidebar/unscoped";
-import {
-	Breadcrumb,
-	BreadcrumbItem,
-	BreadcrumbLink,
-	BreadcrumbList,
-	BreadcrumbPage,
-	BreadcrumbSeparator,
-} from "@/components/ui/breadcrumb";
-import { Separator } from "@/components/ui/separator";
-import {
-	SidebarInset,
-	SidebarProvider,
-	SidebarTrigger,
-} from "@/components/ui/sidebar";
-import { createLazyFileRoute, Outlet } from "@tanstack/react-router";
-
-export const Route = createLazyFileRoute("/(private)/_unscoped")({
-	component: RouteComponent,
-});
-
-function RouteComponent() {
-	return (
-		<SidebarProvider>
-			<UnscopedSidebar />
-			<SidebarInset className="overflow-x-hidden">
-				<header className="flex h-16 shrink-0 items-center gap-2">
-					<div className="flex items-center gap-2 px-4">
-						<SidebarTrigger className="-ml-1" />
-						<Separator orientation="vertical" className="mr-2 h-4" />
-						<Breadcrumb>
-							<BreadcrumbList>
-								<BreadcrumbItem className="hidden md:block">
-									<BreadcrumbLink href="#">
-										Building Your Application
-									</BreadcrumbLink>
-								</BreadcrumbItem>
-								<BreadcrumbSeparator className="hidden md:block" />
-								<BreadcrumbItem>
-									<BreadcrumbPage>Data Fetching</BreadcrumbPage>
-								</BreadcrumbItem>
-							</BreadcrumbList>
-						</Breadcrumb>
-					</div>
-				</header>
-				<div className="flex flex-1 flex-col gap-4 p-4 pt-0">
-					<Outlet />
-				</div>
-			</SidebarInset>
-		</SidebarProvider>
-	);
-}
