+++ src/routes/(private)/_unscoped/components/$component_id/index.tsx
@@ -1,9 +1,9 @@
 import { componentQueries } from "@/data/components";
 import { useSuspenseQuery } from "@tanstack/react-query";
-import { createLazyFileRoute } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { ComponentGeneralTab } from "@/features/components/detail/general-tab";
 
-export const Route = createLazyFileRoute(
+export const Route = createFileRoute(
 	"/(private)/_unscoped/components/$component_id/"
 )({
 	component: RouteComponent,
