+++ src/routes/(private)/_unscoped/components/$component_id/stacks.tsx
@@ -1,5 +1,6 @@
 import { stackQueries } from "@/data/stacks";
 import { createFileRoute } from "@tanstack/react-router";
+import { ComponentStacksTab } from "@/features/components/detail/stacks-tab";
 
 export const Route = createFileRoute(
 	"/(private)/_unscoped/components/$component_id/stacks"
@@ -7,4 +8,11 @@ export const Route = createFileRoute(
 	loader: ({ context: { queryClient }, params: { component_id } }) => {
 		return queryClient.ensureQueryData(stackQueries.list({ component_id }));
 	},
+	component: RouteComponent,
 });
+
+function RouteComponent() {
+	const { component_id } = Route.useParams();
+
+	return <ComponentStacksTab componentId={component_id} />;
+}
