--- src/routes/(public)/_grid-layout/login.lazy.tsx
@@ -1,27 +1,0 @@
-import {
-	Card,
-	CardContent,
-	CardDescription,
-	CardHeader,
-	CardTitle,
-} from "@/components/ui/card";
-import { LoginForm } from "@/features/login/login-form";
-import { createLazyFileRoute } from "@tanstack/react-router";
-
-export const Route = createLazyFileRoute("/(public)/_grid-layout/login")({
-	component: RouteComponent,
-});
-
-function RouteComponent() {
-	return (
-		<Card className="mx-auto md:w-[400px]">
-			<CardHeader>
-				<CardTitle className="text-2xl">Login</CardTitle>
-				<CardDescription>Use your credentials to log in.</CardDescription>
-			</CardHeader>
-			<CardContent>
-				<LoginForm />
-			</CardContent>
-		</Card>
-	);
-}
